fix: add migration for by_key index
